python-app/requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.3
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fabric==2.6.0
Faker==13.3.0
flake8==4.0.1
flower==1.0.0
humanize==4.0.0
iniconfig==1.1.1
invoke==1.6.0
isort==5.10.1
kombu==5.2.3
mccabe==0.6.1
packaging==21.3
paramiko==2.9.2
pathlib2==2.3.7.post1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
